refactor(db): extract connector value types + retry policy to _connector_types
Moves the database-connector data contract and retry policy out of
amx/db/connector.py into a focused amx/db/_connector_types.py:

- ProfilingError exception
- AssetKind enum (TABLE / VIEW / MATERIALIZED_VIEW / SCHEMA / DATABASE)
- ColumnProfile, TableProfile, AnalyticsMetadata dataclasses
- ConnectionTestResult dataclass
- Retry constants (MAX_CONNECTION_RETRIES, CONNECTION_RETRY_BACKOFF_SEC)
- Transient / non-transient pattern tables and the
  _is_transient_db_connection_error classifier

connector.py re-exports every public name so the ~90 external import
sites (core/inference, core/metadata, agents, CLI, web routers, the
adapter layer, tests) keep working byte-compatible.

connector.py: 1966 -> 1794 LOC (-172, -9%).
